[Fix #134055407] Fix folder structure and app arrangement
