+++ adventures/bucketlist/admin.py
@@ -1,0 +1,6 @@
+from django.contrib import admin
+
+from bucketlist.models import Bucketlist, Item
+
+
+admin.site.register((Bucketlist, Item))
