+++ adventures/bucketlist/models.py
@@ -33,3 +33,22 @@ class Bucketlist(TimeStampMixin):
 
     class Meta:
         ordering = ['name']
+
+
+class Item(TimeStampMixin):
+    """
+    This class contains the database schema of the Items
+    i.e. Table and Columns"""
+
+    name = models.CharField(max_length=100, null=False, unique=True)
+    description = models.TextField(max_length=1000)
+    completed = models.BooleanField(default=False)
+    bucketlist = models.ForeignKey(
+        Bucketlist,
+        on_delete=models.CASCADE,
+        related_name='bucketlist_name')
+
+    def __str__(self):
+        return "{} item added to {} bucketlist".format(
+            self.name,
+            self.bucketlist.name)
