--- adventures/bucketlists/admin.py
@@ -1,5 +1,0 @@
-from django.contrib import admin
-from adventures.items.models import Bucketlist
-
-# Register your models here.
-admin.site.register((Bucketlist))
