--- adventures/items/admin.py
@@ -1,5 +1,0 @@
-from django.contrib import admin
-from adventures.items.models import Item
-
-# Register your models here.
-admin.site.register((Item))
