+++ requirements.txt
@@ -1,13 +1,5 @@
-coverage==4.2
-coveralls==1.1
 Django==1.10.3
-djangorestframework==3.5.2
-passlib==1.6.5
+django-filter==0.15.3
+djangorestframework==3.5.3
+Markdown==2.6.7
 psycopg2==2.6.2
-py==1.4.31
-requests==2.11.1
-six==1.10.0
-SQLAlchemy==1.1.3
-SQLAlchemy-Utils==0.32.9
-tox==2.4.1
-virtualenv==15.0.3
